product float no decimal

+++ models/product_model.py
@@ -1,6 +1,5 @@
 from pydantic import BaseModel, Field
 from datetime import datetime
-from decimal import Decimal
 
 
 class Product(BaseModel):
@@ -9,8 +8,8 @@ class Product(BaseModel):
   code: str = Field(..., max_length=8)
   description: str = Field(...)
   categories: list[str] = Field(...)
-  purchase_price: Decimal = Field(..., decimal_places=2)
-  sale_price: Decimal = Field(..., decimal_places=2)
+  purchase_price: float = Field(...)
+  sale_price: float = Field(...)
   stock: int = Field(default=0)
   created_at: datetime = None
   updated_at: datetime = None
